+++ src/main/java/com/example/onlineexamdemo/infrastructure/po/ExamPaperInfo.java
@@ -37,7 +37,7 @@ public class ExamPaperInfo {
 	@TableField
 	private int examPaperEasy;
 
-	@TableField
+	//@TableField
 	private GradeInfo grade;
 
 }
