+++ src/main/java/com/example/onlineexamdemo/infrastructure/xml/CourseInfoMapper.xml
@@ -3,13 +3,13 @@
 PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.onlineexamdemo.infrastructure.po.CourseInfoMapper">
+<mapper namespace="com.example.onlineexamdemo.infrastructure.mapper.CourseInfoMapper">
 	
 	<resultMap type="com.example.onlineexamdemo.infrastructure.po.CourseInfo" id="queryCourses">
 		<id column="courseId" property="courseId"/>
 		<result column="courseName" property="courseName"/>
 		<result column="division" property="division"/>
-		<association property="grade" javaType="com.example.online.exam.po.GradeInfo">
+		<association property="grade" javaType="com.example.onlineexamdemo.infrastructure.po.GradeInfo">
 			<id column="gradeId" property="gradeId"/>
 			<result column="gradeName" property="gradeName"/>
 		</association>
