+++ src/main/java/com/example/onlineexamdemo/infrastructure/xml/ExamHistoryInfoMapper.xml
@@ -18,7 +18,7 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 	</resultMap>
 	
 	<!-- 查询考试历史信息，针对前台学生查询 -->
-	<select id="getExamHistoryToStudent" parameterType="int" resultType="ExamHistoryPaper">
+	<select id="getExamHistoryToStudent" parameterType="int" resultType="com.example.onlineexamdemo.infrastructure.po.ExamHistoryPaper">
 		SELECT
 		a.historyId,a.examScore,b.examPaperId,b.examPaperName,b.examPaperScore,b.subjectNum,c.beginTime
 		FROM ExamHistoryInfo a
@@ -42,13 +42,13 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 	<resultMap type="com.example.onlineexamdemo.infrastructure.po.ExamHistoryInfo" id="queryExamHistoryToTeacherResultMap">
 		<id column="historyId" property="historyId"/>
 		<result column="examScore" property="examScore"/>
-		<association property="examPaper" javaType="com.taohan.online.exam.po.ExamPaperInfo">
+		<association property="examPaper" javaType="com.example.onlineexamdemo.infrastructure.po.ExamPaperInfo">
 			<id column="examPaperId" property="examPaperId"/>
 			<result column="examPaperName" property="examPaperName"/>
 			<result column="examPaperScore" property="examPaperScore"/>
 			<result column="subjectNum" property="subjectNum"/>
 		</association>
-		<association property="student" javaType="com.taohan.online.exam.po.StudentInfo">
+		<association property="student" javaType="com.example.onlineexamdemo.infrastructure.po.StudentInfo">
 			<id column="studentId" property="studentId"/>
 			<result column="studentName" property="studentName"/>
 		</association>
