+++ src/main/java/com/example/onlineexamdemo/infrastructure/xml/StudentExamInfoMapper.xml
@@ -4,7 +4,7 @@ PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
 
-<mapper namespace="com.example.onlineexamdemo.infrastructure.po.StudentExamInfoMapper">
+<mapper namespace="com.example.onlineexamdemo.infrastructure.mapper.StudentExamInfoMapper">
 
 	<resultMap type="com.example.onlineexamdemo.infrastructure.po.StudentExamInfo" id="baseQueryResultMap">
 		<result column="studentId" property="studentId" />
