+++ src/main/java/com/example/onlineexamdemo/infrastructure/xml/StudentInfoMapper.xml
@@ -3,7 +3,7 @@
 PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.onlineexamdemo.infrastructure.po.StudentInfoMapper">
+<mapper namespace="com.example.onlineexamdemo.infrastructure.mapper.StudentInfoMapper">
 	
 	<resultMap type="com.example.onlineexamdemo.infrastructure.po.StudentInfo" id="queryStudent">
 		<id column="studentId" property="studentId"/>
