+++ src/main/java/com/example/onlineexamdemo/infrastructure/xml/SubjectInfoMapper.xml
@@ -3,7 +3,7 @@
 PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
 "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.example.onlineexamdemo.infrastructure.po.SubjectInfoMapper">
+<mapper namespace="com.example.onlineexamdemo.infrastructure.mapper.SubjectInfoMapper">
 	
 	<resultMap type="com.example.onlineexamdemo.infrastructure.po.SubjectInfo" id="querySubjects">
 		<id column="subjectId" property="subjectId"/>
