+++ src/main/resources/application-test.yml
@@ -3,10 +3,16 @@
 #  服务器启动的端口号
 server:
   port: 8080
+  servlet:
+    context-path: /onlineexamdemo
+    jsp:
+      init-parameters:
+        development: true
+
 
 spring:
   application:
-    name: online_exam-backend
+    name: onlineexamdemo
 
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
@@ -45,15 +51,22 @@ spring:
       allow_unquoted_control_chars: true
       allow_single_quotes: true
 
+
   #  下面这个mvc和web配置，是为了使请求不存在的url时，能够按照通用返回模板返回（在全局异常处理器捕获NoHandlerFoundException即可）
+  view:
+    prefix: admin/
+    suffix: .jsp
   mvc:
     view:
-      prefix: /admin/
+      prefix: admin/
       suffix: .jsp
-    throw-exception-if-no-handler-found: true
+    #throw-exception-if-no-handler-found: true
   web:
     resources:
-      add-mappings: false
+      add-mappings: true
+
+  thymeleaf:
+    enabled: false
 
 #  mapper对应的xml文件存放路径，开启 mybatis 下划线命名法自动到驼峰命名法映射
 mybatis-plus:
